refactor(cli): extract skills command subsystem module (#172)
